bot/requirements.txt
python-telegram-bot==22.5
google-genai>=1.33.0
supabase>=2.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
